--- a/src/Bessels.jl
+++ b/src/Bessels.jl
@@ -31,7 +31,6 @@ include("Float128/besselk.jl")
 include("Float128/bessely.jl")
 include("Float128/constants.jl")
 
-include("chebyshev.jl")
 include("math_constants.jl")
 #include("parse.jl")
 #include("hankel.jl")