--- a/src/besselk.jl
+++ b/src/besselk.jl
@@ -1,16 +1,50 @@
-# Modified Bessel functions of the second kind of order zero and one
-# besselk0, besselk1
-# Scaled modified Bessel functions of the second kind of order zero and one
-# besselk0x, besselk1x
+#    Modified Bessel functions of the second kind of order zero and one
+#                       besselk0, besselk1
+#
+#    Scaled modified Bessel functions of the second kind of order zero and one
+#                       besselk0x, besselk1x
+#
+#    (Scaled) Modified Bessel functions of the second kind of order nu
+#                       besselk, besselkx
+#
+#
+#    Calculation of besselk0 is done in two branches using polynomial approximations [1]
+#
+#    Branch 1: x < 1.0 
+#              besselk0(x) + log(x)*besseli0(x) = P7(x^2)
+#                            besseli0(x) = [x/2]^2 * P6([x/2]^2) + 1
+#    Branch 2: x >= 1.0
+#              sqrt(x) * exp(x) * besselk0(x) = P22(1/x) / Q2(1/x)
+#    where P7, P6, P22, and Q2 are 7, 6, 22, and 2 degree polynomials respectively.
+#
+#
+#
+#    Calculation of besselk1 is done in two branches using polynomial approximations [2]
+#
+#    Branch 1: x < 1.0 
+#              besselk1(x) - log(x)*besseli1(x) - 1/x = x*P8(x^2)
+#                            besseli1(x) = [x/2]^2 * (1 + 0.5 * (*x/2)^2 + (x/2)^4 * P5([x/2]^2))
+#    Branch 2: x >= 1.0
+#              sqrt(x) * exp(x) * besselk1(x) = P22(1/x) / Q2(1/x)
+#    where P8, P5, P22, and Q2 are 8, 5, 22, and 2 degree polynomials respectively.
+#
+#
+#    The polynomial coefficients are taken from boost math library [3].
+#    Evaluation of the coefficients using Remez.jl is prohibitive due to the increase
+#    precision required in ArbNumerics.jl. 
+#
+#    Horner's scheme is then used to evaluate all polynomials.
+# 
+# [1] "Rational Approximations for the Modified Bessel Function of the Second Kind 
+#     - K0(x) for Computations with Double Precision" by Pavel Holoborodko     
+# [2] "Rational Approximations for the Modified Bessel Function of the Second Kind 
+#     - K1(x) for Computations with Double Precision" by Pavel Holoborodko
+# [3] https://github.com/boostorg/math/tree/develop/include/boost/math/special_functions/detail
+"""
+    besselk0(x::T) where T <: Union{Float32, Float64}
 
-#=
-Approximate forms used here are given in
-"Rational Approximations for the Modified Bessel Function of the Second Kind - K0(x) for Computations with Double Precision"
-by Pavel Holoborodko, 
-https://www.advanpix.com/2015/11/25/rational-approximations-for-the-modified-bessel-function-of-the-second-kind-k0-for-computations-with-double-precision/
-Actual coefficients used are from the boost math library.
-https://github.com/boostorg/math/tree/develop/include/boost/math/special_functions/detail
-=#
+Modified Bessel function of the second kind of order zero, ``K_0(x)``.
+"""
 function besselk0(x::T) where T <: Union{Float32, Float64}
     x <= zero(T) && return throw(DomainError(x, "`x` must be nonnegative."))
     if x <= one(T)
@@ -24,6 +58,11 @@ function besselk0(x::T) where T <: Union{Float32, Float64}
         return a * s
     end
 end
+"""
+    besselk0x(x::T) where T <: Union{Float32, Float64}
+
+Scaled modified Bessel function of the second kind of order zero, ``K_0(x)*e^{x}``.
+"""
 function besselk0x(x::T) where T <: Union{Float32, Float64}
     x <= zero(T) && return throw(DomainError(x, "`x` must be nonnegative."))
     if x <= one(T)
@@ -35,6 +74,11 @@ function besselk0x(x::T) where T <: Union{Float32, Float64}
         return muladd(evalpoly(inv(x), P3_k0(T)), inv(evalpoly(inv(x), Q3_k0(T))), one(T)) / sqrt(x)
     end
 end
+"""
+    besselk1(x::T) where T <: Union{Float32, Float64}
+
+Modified Bessel function of the second kind of order one, ``K_1(x)``.
+"""
 function besselk1(x::T) where T <: Union{Float32, Float64}
     x <= zero(T) && return throw(DomainError(x, "`x` must be nonnegative."))
     if x <= one(T)
@@ -51,6 +95,11 @@ function besselk1(x::T) where T <: Union{Float32, Float64}
         return a * s
     end
 end
+"""
+    besselk1x(x::T) where T <: Union{Float32, Float64}
+
+Scaled modified Bessel function of the second kind of order one, ``K_1(x)*e^{x}``.
+"""
 function besselk1x(x::T) where T <: Union{Float32, Float64}
     x <= zero(T) && return throw(DomainError(x, "`x` must be nonnegative."))
     if x <= one(T)