Implementation of `besselk0`, `besselk1` and their scaled versions based on approximations given [here](https://www.advanpix.com/2015/11/25/rational-approximations-for-the-modified-bessel-function-of-the-second-kind-k0-for-computations-with-double-precision/)

The implementation follows similar discussion in #6 by reducing a couple of branches and merging `Float32` and `Float64`.

This comes with quite a performance boost...
```julia
# old version
julia> @benchmark besselk0(x) setup=(x = 100*rand())
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  15.514 ns … 52.198 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     36.564 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   36.300 ns ±  2.779 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▂                                                    █   ▃ ▁
  ▅█▃▁▁▄▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▄▁▄█ █
  15.5 ns      Histogram: log(frequency) by time      38.3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# new version
julia> @benchmark besselk0(x) setup=(x = 100*rand())
BenchmarkTools.Trial: 10000 samples with 999 evaluations.
 Range (min … max):   8.925 ns … 29.659 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     13.693 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   13.692 ns ±  0.867 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                                             ▆█               ▁
  ▅▁▆█▃▃▃▁▁▁▁▁▁▁▁▁▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃███▃▁▁▁▁▁▁▁▁▁▁▅█ █
  8.92 ns      Histogram: log(frequency) by time      15.2 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# comparing to SpecialFunctions.jl
julia> @benchmark besselk(0, x) setup=(x = 100*rand())
BenchmarkTools.Trial: 10000 samples with 815 evaluations.
 Range (min … max):  122.103 ns … 592.071 ns  ┊ GC (min … max): 0.00% … 65.17%
 Time  (median):     200.658 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   215.430 ns ±  46.820 ns  ┊ GC (mean ± σ):  0.04% ±  0.94%

             ▇▆██▆▇▆▅▆▅▄▄▃▃▃▃▂▂▁▁▁▁                             ▂
  ▂▄▅▅▂▆▆▇▇▇▇█████████████████████████▇▇██▇▇▆▇▇▆▇▅▆▆▅▆▇▇▆▅▆▆▆▅▅ █
  122 ns        Histogram: log(frequency) by time        435 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

These implementations also handle large arguments better at the edges of float precision
```julia
# SpecialFunctions.jl
julia> besselk(0, 700.0)
0.0

# Bessels.jl
julia> besselk0(700.0)
4.669776431685378e-306
```