@oscardssmith 2022-01-15
How do the maximum errors compare?