@heltonmc 2022-01-16
Old errors compared to `ArbNumerics.jl`
![Screenshot from 2022-01-16 14-09-09](https://user-images.githubusercontent.com/57471570/149674381-a08b19f8-4b0a-4e92-8528-9700d84c478d.png)

New errors

![Screenshot from 2022-01-16 14-06-47](https://user-images.githubusercontent.com/57471570/149674401-3d90360c-8d85-4d2f-87bb-429a1f8e52bc.png)

New error average rel. error / eps(Float64) = 0.51424
maximum = 2.1939




Optimizing these polynomials any further is pretty prohibitive due to the issue with `besselk` in `ArbNumerics.jl`. I've commented https://github.com/JeffreySarnoff/ArbNumerics.jl/issues/48. This essentially makes `Remez.jl` take hours to give meaningful results after setting precision above 2000.