@heltonmc 2022-01-16
I've added implementations for `besselk(nu,x)` and `besselkx(nu, x)`. These are accurately and quickly calculated with a three term recurrence using `besselk0` and `besselk1`. For small to moderate orders this is ok and most likely the fastest way to do this. 

```julia
# Bessels.jl
julia> @benchmark besselk(50, x) setup=(x=rand()*30)
BenchmarkTools.Trial: 10000 samples with 978 evaluations.
 Range (min … max):  66.969 ns … 108.788 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     72.463 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   72.645 ns ±   1.870 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

         ▃▁                      ▁█▆▆▂     ▃▁▁                 ▂
  ▆▆▃▁▁▁███▆▄▁▃▁▁▅▅▅▁▃▁▁▁▁▄▁▁▁▃▁▃██████▅▅▃████▇▇▆▆▆▆▇██▇▇▆▆▄▆▆ █
  67 ns         Histogram: log(frequency) by time        77 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

## Compared to SpecialFunctions.jl
julia> @benchmark besselk(50, x) setup=(x=rand()*30)
BenchmarkTools.Trial: 10000 samples with 200 evaluations.
 Range (min … max):  399.605 ns … 949.175 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     515.200 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   538.276 ns ±  65.815 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

          ▁▁▁▁  ▅█▇▇▆▅▅▄▄▄▃▄▃▃▃▂▃▂▂▂▁▂▁▁▁▁▁▁ ▁                  ▂
  ▅▆▇██████████████████████████████████████████████▇▇▇▇▇▇█▇▇▆▇▇ █
  400 ns        Histogram: log(frequency) by time        792 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

When the order is large and x is large it is much faster to use a uniform expansion to calculate this. I have added notes in the document about how to implement this. We should come back at a later time to implement this as it requires other bessel functions we haven't implemented yet. 

```julia
# Bessels.jl
julia> @benchmark besselk(120, x) setup=(x=rand()*1000 + 10.0)
BenchmarkTools.Trial: 10000 samples with 995 evaluations.
 Range (min … max):   27.027 ns …   5.351 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     137.620 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   288.813 ns ± 879.224 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆█                                                          ▂ ▁
  ██▇▃▁▁▁▁▁▁▃▃▁▁▁▃▄▃▁▁▄▃▄▄▁▄▄▅▁▄▄▃▅▃▁▄▄▄▄▄▄▅▄▃▄▄▄▄▅▄▅▅▅▄▅▅▄▄▄▅█ █
  27 ns         Histogram: log(frequency) by time       4.79 μs <

 Memory estimate: 0 bytes, allocs estimate: 0.

# SpecialFunctions.jl
julia> @benchmark besselk(120, x) setup=(x=rand()*1000 + 10.0)
BenchmarkTools.Trial: 10000 samples with 631 evaluations.
 Range (min … max):  194.799 ns …  1.524 μs  ┊ GC (min … max): 0.00% … 86.15%
 Time  (median):     207.323 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   206.982 ns ± 19.862 ns  ┊ GC (mean ± σ):  0.13% ±  1.22%

   █                   █▇                                       
  ▄█▂▂▂▃▅▅▂▃▅▃▃▃▃▂▂▂▂▂▃██▇▆▂▅▇▆▄▇▅▄▄▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▂▃▂▁▁▂▂▂▂ ▃
  195 ns          Histogram: frequency by time          229 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

It looks like `SpecialFunctions.jl` switches over to this expansion with the amos routines. Which becomes faster starting at around order 200.

@oscardssmith Could you review this three term recurence function? Its speed is pretty important for fast generation of these higher order terms. It will also be used for the `bessely(nu, x)` implementation. 

One thing we should consider is returning an array of nu values. This is usually how you need them anyway so we should also have an implementation that returns the array.... will have to think about best way to do this.