@heltonmc 2022-01-17
Ok I am going to merge this and open a new PR with the other method I am working on for higher orders. It is very accurate even for lowish orders. We can move the discussion of the recurrence errors there as well to decide how we want to deal with those. We will have to determine the right asymptotic cutoff but I think this discussion will be best as its own PR. It works for both `besseli(nu, x)` and `besselk(nu, x)`